pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "langchain-parallel"
version = "0.1.0"
description = "A LangChain integration for Parallel Web AI services, including Chat and Search."
authors = ["Parallel Team <support@parallel.ai>"]
readme = "README.md"
repository = "https://github.com/parallel-web/langchain-parallel"
license = "MIT"

[tool.mypy]
disallow_untyped_defs = true

# Ignore type errors from the Parallel SDK
[[tool.mypy.overrides]]
module = "parallel.*"
ignore_missing_imports = true
ignore_errors = true

# Relax type checking for our SDK client wrapper
[[tool.mypy.overrides]]
module = "langchain_parallel._client"
disable_error_code = ["arg-type"]

# Relax type checking for extract tool's SDK interaction
[[tool.mypy.overrides]]
module = "langchain_parallel.extract_tool"
disable_error_code = ["assignment"]

# Allow string to SecretStr in integration tests
[[tool.mypy.overrides]]
module = "tests.integration_tests.*"
disable_error_code = ["arg-type"]

[tool.poetry.urls]
"Source Code" = "https://github.com/parallel-web/langchain-parallel"
"Release Notes" = "https://github.com/parallel-web/langchain-parallel/releases"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain-core = "^0.3.76"
openai = "^1.88.0"
pydantic = "^2.11.7"
httpx = "^0.27.0"
parallel-web = "^0.3.3"

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins
    "ASYNC",  # flake8-async
    "C4",     # flake8-comprehensions
    "COM",    # flake8-commas
    "D",      # pydocstyle
    "E",      # pycodestyle error
    "EM",     # flake8-errmsg
    "F",      # pyflakes
    "FA",     # flake8-future-annotations
    "FBT",    # flake8-boolean-trap
    "FLY",    # flake8-flynt
    "I",      # isort
    "ICN",    # flake8-import-conventions
    "INT",    # flake8-gettext
    "ISC",    # isort-comprehensions
    "PGH",    # pygrep-hooks
    "PIE",    # flake8-pie
    "PERF",   # flake8-perf
    "PYI",    # flake8-pyi
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "RSE",    # flake8-rst-docstrings
    "RUF",    # ruff
    "S",      # flake8-bandit
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "SIM",    # flake8-simplify
    "T10",    # flake8-debugger
    "T20",    # flake8-print
    "TID",    # flake8-tidy-imports
    "UP",     # pyupgrade
    "W",      # pycodestyle warning
    "YTT",    # flake8-2020
]
ignore = [
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D104",    # Missing docstring in public package
    "D105",    # Missing docstring in magic method
    "D107",    # Missing docstring in __init__
    "D203",    # Incompatible with D211 (blank line before class)
    "D213",    # Incompatible with D212 (multi-line summary)
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "PERF203", # Rarely useful
    "S112",    # Rarely useful
    "RUF012",  # Doesn't play well with Pydantic
    "SLF001",  # Private member access
    "UP007",   # pyupgrade: non-pep604-annotation-union
    "UP045",   # pyupgrade: non-pep604-annotation-optional
]
unfixable = ["B028"] # People should intentionally tune the stacklevel

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    "S101", # Tests need assertions
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]
"examples/*.py" = ["T201"] # Allow print statements in examples
"docs/*.ipynb" = ["T201", "E501"] # Allow print statements and long lines in documentation
"scripts/*.py" = ["SIM105", "S110"] # Allow try-except-pass in scripts
"langchain_parallel/extract_tool.py" = [
    "FBT001", # Boolean-typed positional argument (matches Extract API design)
    "FBT002", # Boolean default positional argument (matches Extract API design)
]
"langchain_parallel/_client.py" = [
    "FBT001", # Boolean-typed positional argument (matches Extract API design)
]

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --durations=5"
markers = [
    "compile: mark placeholder test used to compile integration tests without running them",
]
asyncio_mode = "auto"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.codespell]
optional = true

[tool.poetry.group.test_integration]
optional = true

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-socket = "^0.7.0"
pytest-watcher = "^0.3.4"
langchain-tests = "^0.3.8"
pytest-mock = "^3.14.1"
freezegun = "^1.2.2"
syrupy = "^4.0.2"

[tool.poetry.group.codespell.dependencies]
codespell = "^2.2.6"

[tool.poetry.group.test_integration.dependencies]

[tool.poetry.group.lint.dependencies]
ruff = "<0.13,>=0.12.2"

[tool.poetry.group.typing.dependencies]
mypy = "^1.10"
types-requests = "^2.31.0"
langchain-core = "^0.3.76"
